services/retriever_drafter/requirements.txt
fastapi
uvicorn
pydantic

google-cloud-bigquery
google-cloud-aiplatform
numpy
